resolved, numpy from 1.18.1 to 1.17.4, use vision 040

+++ references040/detection/coco_eval.py
@@ -109,7 +109,8 @@ class CocoEvaluator(object):
             labels = prediction["labels"].tolist()
 
             rles = [
-                mask_util.encode(np.array(mask[0, :, :, np.newaxis], order="F"))[0]
+                # mask_util.encode(np.array(mask[0, :, :, np.newaxis], order="F"))[0]
+                mask_util.encode(np.array(mask[0, :, :, np.newaxis], order="F", dtype=np.uint8))[0]
                 for mask in masks
             ]
             for rle in rles:
